Completed draft of run 6 story.
I finished drafting write-up of which genetic operators led to winners, which I think gives us a reasonably complete draft of Section 4, the story of run 6 lexicase, replace-space-with-newline.

There are other parts of that story we could definitely talk about, but I'm not sure there's room. These include:

* The high incidence of cloning
* The collapse of clusters around the important "discovery" in the late 30's/early 40's
* The discovery of the first useful building blocks in the first 2 or three generations

But we should probably move on to some other runs.