Made individual IDs more readable.
David and I both agreed that Neo4J's 5 digit internal IDs don't make great names in the discussion. We haven't yet come up with a clearly better alternative, but as a small step decided to insert a marker character between the first two digits (the generation number) and the last three digits (that individual's position in the generation, which really conveys no information). I've inserted a colon for now, but I've asked David for feedback on dashes and underscores as well.