Removed unneeded node labels.
As per @Vaguery's excellent solution, we've removed the labels in the "big plots" from all the nodes that aren't explicitly discussed in the text. I definitely think this has helped clean up those plots, and allowed us to turn the lexicase plot back to vertical, which is (I think) a good thing.